check_empty_cart

+++ process/xl_order.php
@@ -32,18 +32,27 @@ if (isset($_POST['btnOrder'])) {
     
     }else{
         $check_id = null; }
-    $order = $db->query("INSERT INTO `order` (`fee`,`u_id`,`o_phone`,`o_name`,`adress`,`s_id`,`note`,`suggest`) VALUES ('35000','$u_id','$phone','$name','$address','$check_id','$note','$suggest');");
-    $o_id = $db->query("SELECT * FROM `order` ORDER BY o_id DESC LIMIT 1;")->fetch();
-    $oid = $o_id['o_id'];
     $detail = $db ->query("SELECT * FROM `cart` WHERE `u_id` = $u_id");
-        foreach($detail as $dt){
-            $dp_id =  $dt['p_id'];
-            $amount = $dt['unit'];
-            $detail = $db->query("INSERT INTO `details` (`o_id`,`p_id`,`amount`) VALUES ('$oid','$dp_id','$amount');");
-            $update = $db->query("UPDATE `product` SET remain = remain - $amount WHERE p_id = $dp_id");
-        }
-        $sql = $db->query("DELETE FROM cart WHERE u_id='$u_id'");
+    if($detail->rowCount() == 0){
+        $tb="Please add product in your cart";
+        header("location: ../cart.php?tb= '$tb'");
+    }
+    else{
+        $order = $db->query("INSERT INTO `order` (`fee`,`u_id`,`o_phone`,`o_name`,`adress`,`s_id`,`note`,`suggest`) VALUES ('35000','$u_id','$phone','$name','$address','$check_id','$note','$suggest');");
+        $o_id = $db->query("SELECT * FROM `order` ORDER BY o_id DESC LIMIT 1;")->fetch();
+        $oid = $o_id['o_id'];
+    
+            foreach($detail as $dt){
+                $dp_id =  $dt['p_id'];
+                $amount = $dt['unit'];
+                $detail = $db->query("INSERT INTO `details` (`o_id`,`p_id`,`amount`) VALUES ('$oid','$dp_id','$amount');");
+                $update = $db->query("UPDATE `product` SET remain = remain - $amount WHERE p_id = $dp_id");
+            }
+            $sql = $db->query("DELETE FROM cart WHERE u_id='$u_id'");
+            header("Location:../order.php"); 
+    }
+
 }
-header("Location:../order.php"); 
+
 
 ?>
